+++ NEWS.md
@@ -8,6 +8,8 @@
 * Report parameters can now be passed in with the request body using a POST
   request. The parameters must be formatted as JSON (#27)
 * Report caching can now be scoped to the client id (#31)
+* Quarto reports based on python now works properly with regards to extracting
+  output formats and parameters (#29)
 
 # routr 1.0.0